requirements.txt
Django==1.6.2
MySQL-python==1.2.5
South==0.8.4
django-bootstrap3==4.2.0
django-recaptcha==1.0
django-registration==1.0
wsgiref==0.1.2
